requirements.txt
Django==4.0.5
django-bootstrap3==21.2
channels==3.0.4
channels-redis==3.4.0
psycopg2
autopep8
uvicorn[standard]
django-celery-beat
django-compressor
flower
django-celery-beat
redis
python-dotenv
Pillow
djangorestframework
markdown
django-filter
pycryptodome